refactor: :recycle: DateWrapper 의 폴더, 파일 명 변경, 필요 없어진 함수 삭제
- close #288
