+++ app/src/api/cursusUser/cursusUser.service.ts
@@ -18,11 +18,11 @@ import type {
 import type { IntRecord } from 'src/common/models/common.valueRecord.model';
 import type { UserFullProfile } from 'src/common/userFullProfile';
 import type { DateRange } from 'src/dateRange/dtos/dateRange.dto';
+import { DateWrapper } from 'src/dateWrapper/dateWrapper';
 import type {
   IntPerCircle,
   UserCountPerLevel,
 } from 'src/page/home/user/models/home.user.model';
-import { DateWrapper } from 'src/statDate/StatDate';
 import { CoalitionService } from '../coalition/coalition.service';
 import { lookupCoalition } from '../coalition/db/coalition.database.aggregate';
 import { lookupCoalitionsUser } from '../coalitionsUser/db/coalitionsUser.database.aggregate';
