+++ app/src/lambda/lambda.service.ts
@@ -35,7 +35,7 @@ import { CacheUtilService } from 'src/cache/cache.util.service';
 import type { UserFullProfile } from 'src/common/userFullProfile';
 import { DateRangeService } from 'src/dateRange/dateRange.service';
 import { DateTemplate, type DateRange } from 'src/dateRange/dtos/dateRange.dto';
-import { DateWrapper } from 'src/statDate/StatDate';
+import { DateWrapper } from 'src/dateWrapper/dateWrapper';
 
 type UpdateRankingByDateTemplateFn = (
   userFullProfiles: UserFullProfile[],
